Reusing Navigable mixin
