+++ app/models/concerns/exercise_navigation.rb
@@ -1,7 +1,8 @@
-module OnGuide
+module ExerciseNavigation
   extend ActiveSupport::Concern
 
   included do
+    include Navigable
     include WithParent
 
     belongs_to :guide
